fix: Set `XDG_CURRENT_DESKTOP` to `cosmic` instead of `COSMIC`
`XDG_CURRENT_DESKTOP` environment variable is expected to have a lower
case value. One of purposes of that variable is determining the XDG
portal to run and an upper-case value breaks the mechanism. The main
visible symptome of the incorrect value is broken screen sharing and
cosmic-screenshot showing the following error:

```
failed to send screenshot request: Zbus(MethodError(OwnedErrorName([...]
```

Overwriting that variable with `cosmic` is reported as a workaround by
people facing the issue.[0][1]

The upper-case value was specified in two places:

- `start-cosmic` script, which was explicitly setting the
  `XDG_CURRENT_DESKTOP` variable.
- `cosmic.desktop` entry, used by cosmic-greeter to populate the
  variable.[2]

To make sure the issue is gone, change it to lower case in both places.

Fixes pop-os/xdg-desktop-portal-cosmic#93

[0] https://aur.archlinux.org/packages/cosmic-screenshot-git#comment-980452
[1] https://github.com/pop-os/cosmic-screenshot/issues/3
[2] https://github.com/pop-os/cosmic-greeter/blob/7366808e8ca59c5ba8c4cfdf39dec26e52ca82c1/src/greeter.rs#L242-L247