fix: Allow to build with `autostart` and without `systemd` feature
Before this change, import of `systemd::is_systemd_used` was hidden
behind `cfg(feature = "systemd")` predicate. But that function is
expected to be present in the autostart block even if `systemd` feature
is disabled.

Fix that by importing `is_systemd_used` unconditionally.

Provide an implementation of `is_systemd_used` for
`cfg(not(feature = "systemd"))` which just returns `false`.

Fixes: #136