Revert: session: Add custom dconf profile
This commit reverts 87a0644435145b009ea4c159d3992ff279292424.

Setting a custom profile for the COSMIC session causes that all apps
launched in it save their settings into this custom profile rather than
the into main user profile, so these settings are lost when the user logs
in into a different desktop environment.

Since the custom gtk.css is cleaned up on exit, and the GTK theme override
is an optional feature, it should be safe to use the main dconf profile.